big iovec refactor with NonNull trick
We want to expose a wrapper for OwningIovec that allows mutable
access (to consume bytes and use the underlying ByteArena), but
does not otherwise allow direct mutation of the OwningIovec itself
(i.e, can't push new slices in, manage backreferences, or swap out
the OeningIovec).

The wrapper's operations are covariant in the slice's lifetime, since
it doesn't expose methods to store new slices in the iovec.
Unfortunately, `&mut OwningIovec<'life>` is invariant in `'life`
(the correct call when one could both extract a 'life-slice out of
the OwningIovec and stick a new one in).

Construct a covariant but mutable wrapper by converting `&mut` to
`NonNull` in a no-Clone type that can only be constructed
from a `&mut OwningIovec`.  The wrapper contains a lifetime derived
from the `&mut` reference, so is assumed to extend a mutable reference
to the iovec.  However, it actually contains a NonNull, so
it's allowed to restrict its lifetime covariantly.

NonNull makes sure to support this use case in Miri, by converting
mutable reference to pointer to mut, and only then to pointer to
const (https://github.com/rust-lang/rust/pull/56161/commits/21b5950cf92a978f88dd65f408382637b60c52b5).