Add sliding deque crate
We want a FIFO queue with contiguous elements (OSes work with
arrays of iovecs, not the pair of arrays of iovs we get from
a regular VecDeque).  Waste some space to keep everything
contiguous and only memmove over consumed space periodically,
with a `SlidingDeque`.

We also want a sorted copntainer that's mostly treated like a FIFO
queue with random reads.  We expect that container to mostly be small;
the `SortedDeque` gives us that, on top of `SmallVec`.

Start reading at `sliding_deque`: `sorted_deque` builds on top of it.