Add owning_iovec crate
OwningIovec is a decent chunk of memory management functionality.  The
OwningIovec data structure exists to centralise buffering in one
places.  It can be used in a zero-copy fashion, for input slices that
outlive the `OwningIovec`'s lifetime parameter (which can always be 'static).

When zero-copy isn't possible or wanted, the `OwningIovec` can also
allocate from an arena and maintain ownership of the slices.  We also
detect the common case of pushing new contiguous allocations from the
same arena to the `OwningIovec` and accelerate that path: we avoid
`Arc` increments and merge the slices to we only have to track one.

For producers, the OwningIovec can also manage backpatch locations:
writers may insert *fixed size* placeholders, and populate them after
the fact (which can be useful for things like checksums or size
headers).  It's hard to describe that interface to the borrow checker,
so *programming errors* can turn into panics, but it's probably better
than trying to handle backpatching "by hand."

For readers, the OwningIovec can return the "stable" prefix of
buffered slices: those are the slices that are fully populated,
without any unpopulated placeholder in or before them.  Readers may
consume incrementally from the `OwningIovec` (e.g., to write to a file
while its contents are still being produced); the `OwningIovec`'s read
pointer will move forward, without affecting the writer, and owned
storage be freed, one arena chunk at a time.

There is currently no way to zero-copy move slices from one
OwningIovec to another, but it's probably doable, by passing around
`(Buf<[IoSlice]>, SmallVec<[Anchor; ...]>)` pairs.

The lifetimes here are tricky. Use MIRI and a lot of unit tests.