Add vouched_time crate
In distributed systems like the woodpile, it's important for some
timestamps to be in the same vicinity as the "real" real time.

Unlike Spanner, the error bar on the timestamps doesn't have to be
tight to the microsecond.  However, we assume that all peers can
agreee on the current time up to ~2 seconds (the error budget is
split evenly between leap second handling and clock sync error).

The vouched time crate uses a slowly updated "trusted" time and
vouches that a fresh timestamp isn't too far from the most recent
trusted time.  Roughtime servers were designed as trusted time
sources, but the ecosystem doesn't look great (very few
implementations, the latest version of the RFC is very different from
the most common implementation, prototype-quality Rust client)...
so we just use ctime from trusted mount points (e.g., NFS servers):
on NFS, the ctime is always set by the server, so, while that time
source might be badly synchronised with reality, everyone should
be able to agree on the same time.