vouched_time/atomic_base_time: ensure monotonic updates
If the new base time is earlier than the current snapshot's base
time, skip the update.

The vouched time module is supposed to mostly help against bad
timestamps from the future.  That's moot if we think the base
time can't be trusted and must sometimes be wound back.

What seems more likely is that concurrent updates will compute
a new base time and complete out of order; when that happens,
it makes more sense to keep the highest (most recent) base time
value instead of the last one written.