vouched_time: add first backend, nfs_voucher
We want to use this to coordinate accesses to a shared NFS mount; it
makes sense to assume everyone can access that NFS server.  After that,
it's pretty bad news if your NFS server can't keep track of time
remotely correctly... and maybe we can just assume it's the server
that's correct and everyone else is wrong.