Add trivial_zip_wrapper crate
We want to use the zip file format as a container of raw bytes (in
uncompressed STORED format).  We also want to do that for input files
that live on NFS.  Copying from NFS with a read(2)/write(2) loop is
usually latency-bound, so we instead want to stick the input bytes in
the zip container with `std::io::copy`, which can then use
`copy_file_range` or similar accelerated kernel/server-side copies.

The trivial_zip_wrapper does that, in zip64 format.  Of course, if we
don't want to actually read the data, we can't compute its CRC, so the
checksum fields are always corrupt (0).  The rest of the zip is valid
though, so we can use standard readers as long as we don't check for
corruption or the check happens late enough.  We should even be able to
"fix" the archives and decompress them at the command line.

Some filesystems also benefit from logical alignment.  For example,
ZFS can (in some configurations) use record-level Copy-on-Write when
the source and destinations's logical offsets are aligned to the
record size (128 KB by default).  Entries in a zip file can be
separated by arbitrary data, so the `trivial_zip_wrapper` crate
supports such alignment, by padding *before* the local header (the
payload must always come immediately after the local header).

Compare the code in headers.rs with the quoted sections of the
PKWARE app note, then lib.rs.  The only interesting thing in
lib.rs is the computating to insert padding before the local
file header, to ensure the *payload* is correctly aligned.