Finally add woodpile crate
The woodpile crate exposes a partially ordered log of records,
where records may be transient (i.e., be visible and eventually
go away) until the record is old enough to be "stable."

The logs are internally implemented as single-writer files in
time bucket directories on disk, with NFS-compatible logic.

At a high level, the life cycle of a log record is:

1. Write to a file in the current time bucket for the destination
   log.

2. Readers can now see the new record

3. The time bucket closes, writers are expected to stop writing
   there

4. Enough time has passed that writers should be done writing;
   copy the contents of the log files to a zip snapshot.

5. The contents of the zip snapshot are now frozen and authoritative.

The only reason the zip snapshot might be a subset of what readers
saw on disk is if a reader caught a slow writer while the snapshot
wasn't available yet.

Multi-writer logs should be combined with a lease system that will
detect slow writers and treat them as crashes (i.e., wait until
lease expires), and the leases should be long enough (20+ seconds
since the last refresh) that readers will wait for the snapshot.

See lib.rs for a bunch of time constants and how they relate to
each other.

Then `close.rs` should solidify the idea, especially after also
reading `epoch_writer.rs`.

The write-side then builds on top of that with `shard_writer.rs`.

The read-side dual of `shard_writer` is `shard_reader`, a thin
wrapper around `hcobs::StreamReader`.  We expect to find multiple
log files in a single directory; the `pile_reader` module handles
that.  We also expect reader to "tail" from the same log files
multiple times; `pile` handles the state management to remember
where to resume `pile_reader`s.

Finally, `log.rs` exposes a nice interface to apply `close.rs`,
`shard_writer.rs`, and `pile.rs` to a fixed directory.