hcobs: add noop and hcobs examples
The noop binary reads 512 KB at a time from stdin and spits it
back to stdout.

The hcobs binary either encodes from stdin to stdout (default),
or, if `argv[1]` starts with "d", decodes from stdin to stdout;
in both cases, input is performed in 512 KB chunks.

These examples let us compare the sizes of noop VS hcobs,
to estimate the space overhead of the hcobs library and its
dependencies, and also lets us confirm that the encoding/decoding
both use a bounded amount of buffer space.

On linux/aarch64, the diff between release builds of noop and hcobs
shows:

- .text: + 25 KB
- .rodata: +3596 bytes
- .data.rel.ro: +2928 bytes

And

```
$ echo foo | target/release/examples/hcobs | target/release/examples/hcobs d
foo
$ target/release/examples/hcobs < /dev/zero | target/release/examples/hcobs d > /dev/null  # heap quickly stops growing
...
```